+++ app/controllers/editorial/news_controller.rb
@@ -7,7 +7,11 @@ module Editorial
     end
 
     def show
-      @article = NewsArticle.find_by(slug: params[:article_slug])
+      @article = NewsArticle.find_by(
+        section: params[:section],
+        slug: params[:slug]
+      )
+
     end
 
     def new
@@ -25,7 +29,6 @@ module Editorial
       # TODO: Handle draft submissions with no sections.
       # This could be handled via a user's context under /editorial/submissions.
       if @form.validate(params.require(:node).permit!)
-        @form.parent_id = @news_home.id
         section = Section.find(@form.section_id)
         submission = NodeCreator.new(section, @form).perform!(current_user)
 
@@ -41,10 +44,7 @@ module Editorial
     # A list of sections should be provided to create 'distributions',
     # which should be formed based on the user's membership access.
     def set_sections_by_membership
-      @sections = current_user.member_of_sections.reject do |section|
-        section.news?
-      end
-      @news_home = Section.find_by(name: 'news').home_node
+      @sections = current_user.member_of_sections
     end
   end
 end