+++ app/controllers/nodes_controller.rb
@@ -5,7 +5,7 @@ class NodesController < ApplicationController
   decorates_assigned :menu_nodes, with: NodeDecorator
 
   def show
-    @node = Node.find_by_path! params[:path] || ''
+    @node = Node.published.find_by_path! params[:path] || ''
     raise ActiveRecord::RecordNotFound unless can? :read, @node
     @section = @node.section
     set_menu_nodes