+++ app/models/root_node.rb
@@ -1,3 +1,7 @@
 class RootNode < Node
   validates_absence_of :parent, :slug, :section
+
+  def layout
+    'application'
+  end
 end