+++ config/breadcrumbs.rb
@@ -14,7 +14,7 @@ end
 
 crumb :public_node do |node|
   if node.parent.present?
-    link node.name, nodes_path(path: node.path)
+    link node.name, node.full_path
     parent :public_node, node.parent
   else
     link 'Home', root_path
@@ -90,4 +90,14 @@ end
 crumb :new_editorial_request do |section|
   link 'Request Membership', new_editorial_section_request_path(section)
   parent :editorial_root
+end
+
+crumb :public_news_articles do |section|
+  link 'News', section_news_articles_path(section.home_node.slug)
+  parent :public_node, section.home_node
+end
+
+crumb :public_news_article do |node|
+  link node.name, node.full_path
+  parent :public_news_articles, node.section
 end