+++ spec/fabricators/node.rb
@@ -17,6 +17,7 @@ end
 Fabricator(:news_article, from: :node, class_name: :news_article) do
   name { Fabricate.sequence(:news_name) { |i| "news-#{i}" } }
   short_summary 'foo'
+  release_date Date.today
   summary 'foobar'
 end